+++ rust/main/agents/relayer/src/test_utils/dummy_data.rs
@@ -76,6 +76,7 @@ pub fn dummy_metadata_builder(
         db.clone(),
         IsmAwareAppContextClassifier::new(default_ism_getter.clone(), vec![]),
         IsmCachePolicyClassifier::new(default_ism_getter, Default::default()),
+        None,
     )
 }