+++ typescript/cli/src/tests/warp/warp-check.e2e-test.ts
@@ -2,7 +2,11 @@ import { expect } from 'chai';
 import { Signer, Wallet, ethers } from 'ethers';
 import { zeroAddress } from 'viem';
 
-import { ERC20Test, HypERC20Collateral__factory } from '@hyperlane-xyz/core';
+import {
+  ERC20Test,
+  HypERC20Collateral__factory,
+  Mailbox__factory,
+} from '@hyperlane-xyz/core';
 import {
   ChainAddresses,
   createWarpRouteConfigId,
@@ -372,6 +376,87 @@ describe('hyperlane warp check e2e tests', async function () {
       const expectedActualText = `ACTUAL: ""\n`;
       const expectedDiffText = `      EXPECTED:
         address: "${addressToBytes32(warpCore.tokens[0].addressOrDenom!)}"`;
+
+      const output = await hyperlaneWarpCheckRaw({
+        warpDeployPath: WARP_DEPLOY_OUTPUT_PATH,
+        warpCoreConfigPath: combinedWarpCoreConfigPath,
+      }).nothrow();
+
+      expect(output.exitCode).to.equal(1);
+      expect(output.text()).to.includes(expectedDiffText);
+      expect(output.text()).to.includes(expectedActualText);
+    });
+
+    it(`should find differences in the hook config between the local and on chain config if it needs to be expanded`, async function () {
+      warpConfig[CHAIN_NAME_2].hook = {
+        type: HookType.MERKLE_TREE,
+      };
+
+      const mailboxInstance = Mailbox__factory.connect(
+        chain2Addresses.mailbox,
+        signer,
+      );
+      const hookAddress = await mailboxInstance.callStatic.defaultHook();
+
+      writeYamlOrJson(WARP_DEPLOY_OUTPUT_PATH, warpConfig);
+      writeYamlOrJson(
+        combinedWarpCoreConfigPath.replace('-config.yaml', '-deploy.yaml'),
+        warpConfig,
+      );
+
+      const warpDeployConfig = warpConfig;
+      await hyperlaneWarpDeploy(WARP_DEPLOY_OUTPUT_PATH);
+
+      const expectedOwner = (await signer.getAddress()).toLowerCase();
+      warpDeployConfig[CHAIN_NAME_2].hook = {
+        type: HookType.FALLBACK_ROUTING,
+        domains: {},
+        fallback: hookAddress,
+        owner: expectedOwner,
+      };
+
+      writeYamlOrJson(WARP_DEPLOY_OUTPUT_PATH, warpDeployConfig);
+
+      const expectedActualText = `ACTUAL: ${HookType.MERKLE_TREE}\n`;
+      const expectedDiffText = `EXPECTED: ${HookType.FALLBACK_ROUTING}`;
+
+      const expectedFallbackDiff = `    fallback:
+      ACTUAL: ""
+      EXPECTED:
+        owner: "${expectedOwner}"
+        type: protocolFee
+        maxProtocolFee: "1000000000000000000"
+        protocolFee: "200000000000000"
+        beneficiary: "${expectedOwner}"`;
+
+      const output = await hyperlaneWarpCheckRaw({
+        warpDeployPath: WARP_DEPLOY_OUTPUT_PATH,
+        warpCoreConfigPath: combinedWarpCoreConfigPath,
+      }).nothrow();
+
+      expect(output.exitCode).to.equal(1);
+      expect(output.text()).to.includes(expectedDiffText);
+      expect(output.text()).to.includes(expectedActualText);
+      expect(output.text()).to.includes(expectedFallbackDiff);
+    });
+
+    it(`should find differences in the hook config between the local and on chain config if it compares the hook addresses`, async function () {
+      const mailboxInstance = Mailbox__factory.connect(
+        chain2Addresses.mailbox,
+        signer,
+      );
+      const hookAddress = (
+        await mailboxInstance.callStatic.defaultHook()
+      ).toLowerCase();
+
+      const warpDeployConfig = await deployAndExportWarpRoute();
+      await hyperlaneWarpDeploy(WARP_DEPLOY_OUTPUT_PATH);
+
+      warpDeployConfig[CHAIN_NAME_2].hook = hookAddress;
+      writeYamlOrJson(WARP_DEPLOY_OUTPUT_PATH, warpDeployConfig);
+
+      const expectedActualText = `ACTUAL: "${zeroAddress}"\n`;
+      const expectedDiffText = `EXPECTED: "${hookAddress}"`;
 
       const output = await hyperlaneWarpCheckRaw({
         warpDeployPath: WARP_DEPLOY_OUTPUT_PATH,