+++ typescript/infra/config/environments/mainnet3/owners.ts
@@ -87,12 +87,18 @@ export const chainOwners: ChainMap<OwnableConfig> = {
   milkyway: {
     owner: 'TODO: configure milkyway owner',
   },
+  kyve: {
+    owner: 'TODO: configure kyve owner',
+  },
   soon: {
     // Squads vault
     owner: 'E3QPSn2Upk2EiidSsUqSQpRCc7BhzWZCKpVncemz3p62',
   },
   sonicsvm: {
     // Will move to a Squads once it's live
+    owner: '9bRSUPjfS3xS6n5EfkJzHFTRDa4AHLda8BU2pP4HoWnf',
+  },
+  svmbnb: {
     owner: '9bRSUPjfS3xS6n5EfkJzHFTRDa4AHLda8BU2pP4HoWnf',
   },
 };