+++ typescript/utils/src/addresses.ts
@@ -57,6 +57,10 @@ export function isAddressCosmos(address: Address) {
   );
 }
 
+export function isCosmosIbcDenomAddress(address: Address): boolean {
+  return IBC_DENOM_REGEX.test(address);
+}
+
 export function isAddressStarknet(address: Address) {
   return STARKNET_ADDRESS_REGEX.test(address);
 }