+++ typescript/utils/src/index.ts
@@ -23,6 +23,7 @@ export {
   getAddressProtocolType,
   isAddress,
   isAddressCosmos,
+  isCosmosIbcDenomAddress,
   isAddressEvm,
   isAddressSealevel,
   isAddressStarknet,
@@ -120,6 +121,7 @@ export {
   deepEquals,
   deepFind,
   diffObjMerge,
+  keepOnlyDiffObjects,
   invertKeysAndValues,
   isObjEmpty,
   isObject,
@@ -137,6 +139,7 @@ export {
   stringifyObject,
   transformObj,
   TransformObjectTransformer,
+  sortArraysInObject,
 } from './objects.js';
 export { Result, failure, success } from './result.js';
 export {