+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hyperlane-xyz/utils
 
+## 12.6.0
+
+### Minor Changes
+
+- d182d7d: Adds the sortArraysInObject function to properly sort arrays in an object recursively given an optional sort function
+- b360802: Add the isCosmosIbcDenomAddress function and improve the config expansion logic to correctly format the destination gas
+
 ## 12.5.0
 
 ## 12.4.0